dtv-scan-tables - dvb-s - update and rename the config file for Atlantic-Bird-3-5.0W - Ku Band
Hello,

The latest update for the file Atlantic-Bird-3-5.0W (Ku Band) is 2015-03-28:
https://git.linuxtv.org/dtv-scan-tables.git/log/dvb-s/Atlantic-Bird-3-5.0W
but since March 2015 there are a lot of changes in the tranponders list.
And now, a lot of transponders for Algerian, French and Italian TV and
Radio channels are missing.
Attached is an updated config scan file (Atlantic-Bird-3-5.0W.patch
file) for Atlantic-Bird-3-5.0W transponders in dvbv5 format.
Could you please ensure that this new patch is committed into the DVB
source code tree repositery?
PS : Attached there is also the full up to date file :
Atlantic-Bird-3-5.0W-2017-01-09.NEW.FULL if you need it

Thank you

Best regards