Intelsat34-55.5W: add a Ku-Band transponder
According with http://www.satbeams.com/satellites?id=2641, Intelsat 34
has:

	"22 C-band channels to provide services to to North and South America and
	 Europe and 18 Ku-band channels to provide services to Mexico, Central
	 America, Brazil, the Caribbean, Europe, a portion of the United States,
	 and the northern part of the Atlantic Ocean. It will serve the role
	 originally intended for Intelsat 27 that was lost at launch failure and
	 will replace Intelsat 805 and Galaxy 11 at 304.5 degrees East longitude
	 position."

Currently, we have only C-band frequencies for this satellite.

Add Ku-bands as used currently in South America.

The transponders that were obtained from Vivo TV package
matches the ones at:
	http://www.portaleds.com/satelites/listar.php?sat=3045
with a difference of 1-2 MHz.

The LNBf I have is not capable of tuning the two IP data channels
there, but I'm adding it, anyway, as it could be useful for someone.

Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>