dvb-t: Updated uk-Hannington mux frequencies
I discovered while setting up tvheadend that the Hannington Mux channels/frequencies were out of date.
I've updated the file based on the ukfree.tv, aerialsandtv and main freeview websites, and based on the results from my tvheadend install.
C32 was previously used for HD and is now used as one of the "Local" muxes for Hannington.

References:
https://ukfree.tv/transmitters/tv/Hannington
https://www.aerialsandtv.com/knowledge/transmitters/hannington-transmitter
https://www.freeview.co.uk/help/coverage-checker/detailed-view?postcode=RG265UD

Link: https://github.com/tvheadend/dtv-scan-tables/pull/149
Signed-off-by: Andrew Burns <andrewjburns95@gmail.com>